+++ app/Http/Controllers/TanqueMovimentacaoController.php
@@ -6,6 +6,7 @@ use App\Combustivel;
 use App\TanqueMovimentacao;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Session;
 
 class TanqueMovimentacaoController extends Controller
@@ -21,45 +22,40 @@ class TanqueMovimentacaoController extends Controller
     );
 
     /**
-     * Create a new controller instance.
-     *
-     * @return void
-     */
-     public function __construct()
-     {
-         $this->middleware('auth');
-     }
-
-    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function index(Request $request)
     {
-        if (isset($request->searchField)) {
-            $movimentacoes = DB::table('tanque_movimentacoes')
-                                ->select('tanque_movimentacoes.*', 'tanques.descricao_tanque', 'combustiveis.descricao')
-                                ->join('tanques', 'tanques.id', 'tanque_movimentacoes.tanque_id')
-                                ->join('combustiveis', 'combustiveis.id', 'tanques.combustivel_id')
-                                ->where([
-                                    ['data_hora_abastecimento', '=', $request->searchField],
-                                    ['entrada_combustivel', true]
-                                ])
-                                ->paginate();
-        } else {
-            $movimentacoes = DB::table('tanque_movimentacoes')
-                                ->select('tanque_movimentacoes.*', 'tanques.descricao_tanque', 'combustiveis.descricao' )
-                                ->join('tanques', 'tanques.id', 'tanque_movimentacoes.tanque_id')
-                                ->join('combustiveis', 'combustiveis.id', 'tanques.combustivel_id')
-                                ->where('entrada_combustivel', true)
-                                ->paginate();
-        }
+        if (Auth::user()->canListarTanqueMovimentacao()) {
+            if (isset($request->searchField)) {
+                $movimentacoes = DB::table('tanque_movimentacoes')
+                                    ->select('tanque_movimentacoes.*', 'tanques.descricao_tanque', 'combustiveis.descricao')
+                                    ->join('tanques', 'tanques.id', 'tanque_movimentacoes.tanque_id')
+                                    ->join('combustiveis', 'combustiveis.id', 'tanques.combustivel_id')
+                                    ->where([
+                                        ['data_hora_abastecimento', '=', $request->searchField],
+                                        ['entrada_combustivel', true]
+                                    ])
+                                    ->paginate();
+            } else {
+                $movimentacoes = DB::table('tanque_movimentacoes')
+                                    ->select('tanque_movimentacoes.*', 'tanques.descricao_tanque', 'combustiveis.descricao' )
+                                    ->join('tanques', 'tanques.id', 'tanque_movimentacoes.tanque_id')
+                                    ->join('combustiveis', 'combustiveis.id', 'tanques.combustivel_id')
+                                    ->where('entrada_combustivel', true)
+                                    ->paginate();
+            }
 
-        return View('tanque_movimentacao.index', [
-            'movimentacoes' => $movimentacoes,
-            'fields' => $this->fields
-        ]);
+            return View('tanque_movimentacao.index', [
+                'movimentacoes' => $movimentacoes,
+                'fields' => $this->fields
+            ]);
+        } else {
+            Session::flash('error', __('messages.access_denied'));
+            return redirect()->back();   
+        }
     }
 
     /**
@@ -69,9 +65,14 @@ class TanqueMovimentacaoController extends Controller
      */
     public function create()
     {
-        $combustiveis = Combustivel::where('ativo', true)->get();
+        if (Auth::user()->canCadastrarTanqueMovimentacao()) {
+            $combustiveis = Combustivel::where('ativo', true)->get();
 
-        return View('tanque_movimentacao.create')->withCombustiveis($combustiveis);
+            return View('tanque_movimentacao.create')->withCombustiveis($combustiveis);
+        } else {
+            Session::flash('error', __('messages.access_denied'));
+            return redirect()->back();   
+        }
     }
 
     /**
@@ -82,42 +83,40 @@ class TanqueMovimentacaoController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'data_movimentacao' => 'required',
-            'tanque_id' => 'required',
-            'quantidade_combustivel' => 'required'
-        ]);
+        if (Auth::user()->canCadastrarTanqueMovimentacao()) {
+            $this->validate($request, [
+                'documento' => 'required',
+                'data_movimentacao' => 'required',
+                'tanque_id' => 'required',
+                'quantidade_combustivel' => 'required'
+            ]);
 
-        try {
-            $movimentacao = new TanqueMovimentacao;
-            $movimentacao->data_movimentacao = \DateTime::createFromFormat('d/m/Y', $request->data_movimentacao)->format('Y-m-d');
-            $movimentacao->documento = $request->documento;
-            $movimentacao->tanque_id = $request->tanque_id;
-            $movimentacao->quantidade_combustivel = $request->quantidade_combustivel;
-            $movimentacao->entrada_combustivel = true;
+            try {
+                $movimentacao = new TanqueMovimentacao;
+                $movimentacao->data_movimentacao = \DateTime::createFromFormat('d/m/Y', $request->data_movimentacao)->format('Y-m-d');
+                $movimentacao->documento = $request->documento;
+                $movimentacao->tanque_id = $request->tanque_id;
+                $movimentacao->quantidade_combustivel = $request->quantidade_combustivel;
+                $movimentacao->entrada_combustivel = true;
 
 
-            if ($movimentacao->save()) {
-                Session::flash('success', 'Entrada de Combustível '.$movimentacao->documento.' cadastrada com sucesso.');
-                return redirect()->action('TanqueMovimentacaoController@index');
+                if ($movimentacao->save()) {
+                    Session::flash('success', __('messages.create_success_f', [
+                        'model' => __('models.tanque_movimentacao'),
+                        'name' => $movimentacao->documento
+                    ]));
+                    return redirect()->action('TanqueMovimentacaoController@index');
+                }
+            } catch (\Exception $e) {
+                Session::flash('error', __('messages.exception', [
+                    'exception' => $e->getMessage()
+                ]));
+                return redirect()->back()->withInput();
             }
-        } catch (\Exception $e) {
-            Session::flash('error', 'Ocorreu um erro ao salvar os dados. '.$e->getMessage());
-            return View('tanque_movimentacao.create', [
-                'combustiveis' => Combustivel::where('ativo', true)->get()
-            ]); 
+        } else {
+            Session::flash('error', __('messages.access_denied'));
+            return redirect()->back();   
         }
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\TanqueMovimentacao  $tanqueMovimentacao
-     * @return \Illuminate\Http\Response
-     */
-    public function show(TanqueMovimentacao $tanqueMovimentacao)
-    {
-        //
     }
 
     /**
@@ -128,19 +127,24 @@ class TanqueMovimentacaoController extends Controller
      */
     public function edit(TanqueMovimentacao $tanqueMovimentacao)
     {
-        $movimentacao = TanqueMovimentacao::select('tanque_movimentacoes.*', 'combustiveis.id')
-                                    ->join('tanques', 'tanques.id', 'tanque_movimentacoes.tanque_id')
-                                    ->join('combustiveis', 'combustiveis.id', 'tanques.combustivel_id')
-                                    ->where('tanque_movimentacoes.id', $tanqueMovimentacao->id)
-                                    ->first();        
+        if (Auth::user()->canAlterarTanqueMovimentacao()) {
+            $movimentacao = TanqueMovimentacao::select('tanque_movimentacoes.*', 'combustiveis.id')
+                                        ->join('tanques', 'tanques.id', 'tanque_movimentacoes.tanque_id')
+                                        ->join('combustiveis', 'combustiveis.id', 'tanques.combustivel_id')
+                                        ->where('tanque_movimentacoes.id', $tanqueMovimentacao->id)
+                                        ->first();        
 
-        $combustiveis = Combustivel::select('combustiveis.*')
-                                    ->join('tanques', 'tanques.combustivel_id', 'combustiveis.id')
-                                    ->where('combustiveis.ativo', true)
-                                    ->orWhere('tanques.id', $movimentacao->tanque_id)
-                                    ->get();
+            $combustiveis = Combustivel::select('combustiveis.*')
+                                        ->join('tanques', 'tanques.combustivel_id', 'combustiveis.id')
+                                        ->where('combustiveis.ativo', true)
+                                        ->orWhere('tanques.id', $movimentacao->tanque_id)
+                                        ->get();
 
-        return View('tanque_movimentacao.edit')->withMovimentacao($movimentacao)->withCombustiveis($combustiveis);
+            return View('tanque_movimentacao.edit')->withMovimentacao($movimentacao)->withCombustiveis($combustiveis);
+        } else {
+            Session::flash('error', __('messages.access_denied'));
+            return redirect()->back();   
+        }
     }
 
     /**
@@ -152,31 +156,39 @@ class TanqueMovimentacaoController extends Controller
      */
     public function update(Request $request, TanqueMovimentacao $tanqueMovimentacao)
     {
-        $this->validate($request, [
-            'data_movimentacao' => 'required',
-            'tanque_id' => 'required',
-            'quantidade_combustivel' => 'required'
-        ]);
+        if (Auth::user()->canAlterarTanqueMovimentacao()) {
+            $this->validate($request, [
+                'documento' => 'required',
+                'data_movimentacao' => 'required',
+                'tanque_id' => 'required',
+                'quantidade_combustivel' => 'required'
+            ]);
 
-        try {
-            $movimentacao = TanqueMovimentacao::find($tanqueMovimentacao->id);
-            $movimentacao->data_movimentacao = \DateTime::createFromFormat('d/m/Y', $request->data_movimentacao)->format('Y-m-d');
-            $movimentacao->documento = $request->documento;
-            $movimentacao->tanque_id = $request->tanque_id;
-            $movimentacao->quantidade_combustivel = $request->quantidade_combustivel;
-            $movimentacao->entrada_combustivel = true;
-            $movimentacao->ativo = $request->ativo;
+            try {
+                $movimentacao->data_movimentacao = \DateTime::createFromFormat('d/m/Y', $request->data_movimentacao)->format('Y-m-d');
+                $movimentacao->documento = $request->documento;
+                $movimentacao->tanque_id = $request->tanque_id;
+                $movimentacao->quantidade_combustivel = $request->quantidade_combustivel;
+                $movimentacao->entrada_combustivel = true;
+                $movimentacao->ativo = $request->ativo;
 
 
-            if ($movimentacao->save()) {
-                Session::flash('success', 'Entrada de Combustível '.$movimentacao->documento.' alterada com sucesso.');
-                return redirect()->action('TanqueMovimentacaoController@index');
+                if ($movimentacao->save()) {
+                    Session::flash('success', __('messages.update_success_f', [
+                        'model' => __('models.tanque_movimentacao'),
+                        'name' => $movimentacao->documento
+                    ]));
+                    return redirect()->action('TanqueMovimentacaoController@index');
+                }
+            } catch (\Exception $e) {
+                Session::flash('error', __('messages.exception', [
+                    'exception' => $e->getMessage()
+                ]));
+                return redirect()->back()->withInput(); 
             }
-        } catch (\Exception $e) {
-            Session::flash('error', 'Ocorreu um erro ao salvar os dados. '.$e->getMessage());
-            return View('tanque_movimentacao.create', [
-                'combustiveis' => Combustivel::where('ativo', true)->get()
-            ]); 
+        } else {
+            Session::flash('error', __('messages.access_denied'));
+            return redirect()->back();   
         }
     }
 
@@ -188,16 +200,31 @@ class TanqueMovimentacaoController extends Controller
      */
     public function destroy(TanqueMovimentacao $tanqueMovimentacao)
     {
-        try {
-            $tanqueMovimentacao = TanqueMovimentacao::find($tanqueMovimentacao->id);
-            if ($tanqueMovimentacao->delete()) {
-                Session::flash('success', 'Entrada de Combustível '.$tanqueMovimentacao->id.' removido com sucesso.');
-                
+        if (Auth::user()->canExcluirTanqueMovimentacao()) {
+            try {
+                if ($tanqueMovimentacao->delete()) {
+                    Session::flash('success', __('messages.delete_success_f', [
+                        'model' => __('models.tanque_movimentacao'),
+                        'name' => $tanqueMovimentacao->documento
+                    ]));
+                    return redirect()->action('TanqueMovimentacaoController@index');
+                }
+            } catch (\Exception $e) {
+                switch ($e->getCode()) {
+                    case 23000:
+                        Session::flash('error', __('messages.fk_exception'));
+                        break;
+                    default:
+                        Session::flash('error', __('messages.exception', [
+                            'exception' => $e->getMessage()
+                        ]));
+                        break;
+                }
                 return redirect()->action('TanqueMovimentacaoController@index');
             }
-        } catch (\Exception $e) {
-            Session::flash('error', 'Registro não pode ser excluído. '.$e->getMessage());
-            return redirect()->action('TanqueMovimentacaoController@index');
+        } else {
+            Session::flash('error', __('messages.access_denied'));
+            return redirect()->back();   
         }
     }
 }
